split up mat_plot into package
