+++ autoarray/plot/auto_labels.py
@@ -1,0 +1,10 @@
+class AutoLabels:
+    def __init__(
+        self, title=None, ylabel=None, xlabel=None, legend=None, filename=None
+    ):
+
+        self.title = title
+        self.ylabel = ylabel
+        self.xlabel = xlabel
+        self.legend = legend
+        self.filename = filename
